backend/requirements.txt
fastapi==0.121.1
uvicorn==0.37.0
pydantic==2.12.0
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.31
chromadb==1.1.1
openai==2.3.0
PyMuPDF==1.26.6
python-dotenv==1.1.1
python-multipart==0.0.20
tiktoken==0.8.0
requests==2.32.3
